Laravel Sail

+++ database/migrations/2024_11_26_190349_create_sale_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('sale', function (Blueprint $table) {
             $table->id();
-            $table->double('amount', total: 4, places: 2);
+            $table->decimal('amount', total: 4, places: 2);
             $table->dateTime('sale_date')->useCurrent();
             $table->foreignId('screening_id');
             $table->foreign('screening_id')->references('id')->on('screening');
